Adds explicit list of allowed value types
Explicitly defining a set of data types clarifies some past
misunderstandings about what is allowed in resource and subresource
properties - I'm still finding people who are surprised that they're not
allowed to have an array or object value. It also gives server and
client implementations a common grammar to work with.

In addition, we can define some data types that are higher-level than
what JSON gives us natively, making it easier to express the intention
of a value. For example, having an explicit date-time type which encodes
to a formatted string avoids lots of pitfalls in passing date/time
values.

This list can and probably should be expanded with more specialized
types (like differentiating between integer and decimal types).